requirements.txt
python-dotenv
llama-index
pandas
numpy
openpyxl
transformers
torch
tqdm
networkx
scikit-learn